requirements_bot.txt
python-telegram-bot==20.7
redis==5.0.1
python-dotenv==1.0.1
